Redesigned competition model
