+++ app/static/css/competition_create.css
@@ -1,0 +1,190 @@
+/* Layout */
+.model-editor {
+  padding: 1rem;
+}
+
+.model-editor__header {
+  display: flex;
+  flex-direction: column;
+  gap: 0.25rem;
+  margin-bottom: 1rem;
+}
+
+.model-editor .subtitle {
+  color: #6c757d;
+  margin: 0;
+}
+
+.btn-back {
+  display: inline-block;
+  margin-top: 0.5rem;
+  color: #3498db;
+}
+
+.panel {
+  background: #fff;
+  border: 1px solid #e9ecef;
+  border-radius: 8px;
+  padding: 1rem;
+  margin-bottom: 1rem;
+}
+
+.panel-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 0.5rem;
+}
+
+.stack > * + * {
+  margin-top: 0.75rem;
+}
+
+/* Grid helpers */
+.grid {
+  display: grid;
+  gap: 0.75rem;
+}
+.grid-2 { grid-template-columns: repeat(2, minmax(0, 1fr)); }
+.grid-3 { grid-template-columns: repeat(3, minmax(0, 1fr)); }
+
+@media (max-width: 800px) {
+  .grid-2, .grid-3 { grid-template-columns: 1fr; }
+}
+
+/* Cards */
+.card {
+  border: 1px solid #e9ecef;
+  border-radius: 8px;
+  padding: 0.75rem;
+  background: #fafafa;
+}
+
+.card-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 0.5rem;
+}
+
+.subsection {
+  margin-top: 0.75rem;
+}
+
+.subsection-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 1rem;
+}
+
+/* Form fields */
+.form-field {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 0.5rem;
+}
+.form-field input,
+.form-field select,
+.form-field textarea {
+  border: 1px solid #ced4da;
+  border-radius: 6px;
+  padding: 0.45rem 0.6rem;
+}
+.form-field small {
+  color: #6c757d;
+}
+
+.inline {
+  display: grid;
+  grid-template-columns: 1fr 140px 1fr;
+  gap: 0.5rem;
+}
+
+.checkbox {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.5rem;
+  margin-right: 1rem;
+}
+
+/* Buttons */
+.btn {
+  background: #e9ecef;
+  border: 1px solid #ced4da;
+  color: #333;
+  padding: 0.4rem 0.7rem;
+  border-radius: 6px;
+  cursor: pointer;
+}
+.btn:hover { background: #dee2e6; }
+.btn.primary {
+  background: #3498db;
+  border-color: #2980b9;
+  color: #fff;
+}
+.btn.primary:hover { background: #2980b9; }
+.btn.danger {
+  background: #dc3545;
+  border-color: #c82333;
+  color: #fff;
+}
+.btn.small { padding: 0.25rem 0.5rem; font-size: 0.9rem; }
+
+.actions-right {
+  display: flex;
+  align-items: end;
+  justify-content: end;
+}
+
+/* Status + JSON preview */
+.status-msg {
+  margin-left: 0.75rem;
+  color: #2c3e50;
+}
+
+.json-preview {
+  background: #0f172a;
+  color: #e2e8f0;
+  padding: 0.75rem;
+  border-radius: 8px;
+  overflow: auto;
+  max-height: 380px;
+  font-size: 0.9rem;
+}
+
+/* Panels */
+.actions-panel {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.movement-card {
+  max-width: 100%;
+  margin-bottom: 1rem;
+}
+
+.movement-content {
+  padding: 1rem;
+}
+
+.scoring-grid {
+  display: grid;
+  grid-template-columns: 1fr auto 1fr;
+  gap: 0.5rem;
+  align-items: start;
+}
+
+.metrics-section {
+  margin-top: 1rem;
+  border-top: 1px solid #eee;
+  padding-top: 1rem;
+}
+
+.metrics-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 0.5rem;
+}
