+++ app/templates/admin/includes/_competition_templates.html
@@ -1,0 +1,15 @@
+<template id="tpl-event">
+    <!-- Copy the template content from competition_create.html -->
+</template>
+
+<template id="tpl-movement">
+    <!-- Copy the template content from competition_create.html -->
+</template>
+
+<template id="tpl-metric">
+    <!-- Copy the template content from competition_create.html -->
+</template>
+
+<template id="tpl-group">
+    <!-- Copy the template content from competition_create.html -->
+</template>
